pixci-web/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pillow==10.4.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
